CListBox: remove float type
Is not really required

+++ src/gui/widget/listbox.cpp
@@ -73,10 +73,10 @@ void CListBox::paint()
 	if (sbc < 1)
 		sbc = 1;
 
-	float sbh= (sb- 4)/ sbc;
+	int sbh= (sb- 4)/ sbc;
 	int sbs= (selected/listmaxshow);
 
-	frameBuffer->paintBoxRel(x+ width- 13, ypos+ 2+ int(sbs* sbh) , 11, int(sbh),  COL_MENUCONTENT_PLUS_3);
+	frameBuffer->paintBoxRel(x+ width- 13, ypos+ 2+ sbs* sbh , 11, sbh,  COL_MENUCONTENT_PLUS_3);
 }
 
 void CListBox::paintHead()
